requirements.txt
django==5.0.6
pillow==10.3.0
pyttsx3==2.90
rembg==2.0.57
pydub==0.25.1